Fix issues with Kotlin examples in Antora docs
