+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/features/devservices/testcontainers/atdevelopmenttime/dynamicproperties/MyContainersConfiguration.kt
@@ -13,22 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.boot.docs.features.testcontainers.atdevelopmenttime.devtools
 
-import org.springframework.boot.devtools.restart.RestartScope
+package org.springframework.boot.docs.features.devservices.testcontainers.atdevelopmenttime.dynamicproperties
+
 import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection
 import org.springframework.context.annotation.Bean
+import org.springframework.test.context.DynamicPropertyRegistrar;
 import org.testcontainers.containers.MongoDBContainer
 
 @TestConfiguration(proxyBeanMethods = false)
 class MyContainersConfiguration {
 
 	@Bean
-	@RestartScope
-	@ServiceConnection
 	fun mongoDbContainer(): MongoDBContainer {
 		return MongoDBContainer("mongo:5.0")
+	}
+	
+	@Bean
+	fun mongoDbProperties(container: MongoDBContainer): DynamicPropertyRegistrar {
+		return DynamicPropertyRegistrar { properties ->
+			properties.add("spring.data.mongodb.host") { container.host }
+			properties.add("spring.data.mongodb.port") { container.firstMappedPort }
+		}
 	}
 
 }
