+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/features/devservices/testcontainers/atdevelopmenttime/importingcontainerdeclarations/MyContainers.kt
@@ -14,12 +14,28 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.features.testcontainers.atdevelopmenttime.importingcontainerdeclarations
+package org.springframework.boot.docs.features.devservices.testcontainers.atdevelopmenttime.importingcontainerdeclarations
 
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.boot.testcontainers.context.ImportTestcontainers
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection
+import org.testcontainers.containers.MongoDBContainer
+import org.testcontainers.containers.Neo4jContainer
+import org.testcontainers.junit.jupiter.Container
 
-@TestConfiguration(proxyBeanMethods = false)
-@ImportTestcontainers(MyContainers::class)
-class MyContainersConfiguration
+interface MyContainers {
+
+	companion object {
+
+		@Container
+		@ServiceConnection
+		@JvmField
+		val mongoContainer = MongoDBContainer("mongo:5.0")
+
+		@Container
+		@ServiceConnection
+		@JvmField
+		val neo4jContainer = Neo4jContainer("neo4j:5")
+
+	}
+
+}
