+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/features/devservices/testcontainers/atdevelopmenttime/launch/TestMyApplication.kt
@@ -14,16 +14,11 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.features.testcontainers.atdevelopmenttime.test
+package org.springframework.boot.docs.features.devservices.testcontainers.atdevelopmenttime.launch
 
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.docs.features.springapplication.MyApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class MyApplication
+import org.springframework.boot.fromApplication
 
 fun main(args: Array<String>) {
-	runApplication<MyApplication>(*args)
+	fromApplication<MyApplication>().run(*args)
 }
