+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/features/devservices/testcontainers/atdevelopmenttime/test/MyContainersConfiguration.kt
@@ -14,6 +14,21 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.web.servlet.springmvc.errorhandling.errorpageswithoutspringmvc
+package org.springframework.boot.docs.features.devservices.testcontainers.atdevelopmenttime.test
 
-class SomeController
+import org.springframework.boot.test.context.TestConfiguration
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection
+import org.springframework.context.annotation.Bean
+import org.testcontainers.containers.Neo4jContainer
+
+@TestConfiguration(proxyBeanMethods = false)
+class MyContainersConfiguration {
+
+	@Bean
+	@ServiceConnection
+	fun neo4jContainer(): Neo4jContainer<*> {
+		return Neo4jContainer("neo4j:5")
+	}
+
+}
+
