+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/packaging/nativeimage/advanced/nestedconfigurationproperties/MyPropertiesKotlin.kt
@@ -18,6 +18,7 @@ package org.springframework.boot.docs.packaging.nativeimage.advanced.nestedconfi
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.NestedConfigurationProperty
+import org.springframework.boot.docs.packaging.nativeimage.advanced.nestedconfigurationproperties.nested.Nested
 
 @ConfigurationProperties("my.properties")
 data class MyPropertiesKotlin(
